feat: handling secret dropdown differently for dedicated admin
this PR contains:
different ways of listing and creating secrets of cluster admins and dedicated admins.
the default namespace for cluster-admin is redhat-data-federation.
the default namespace for dedicated admin is dedicated-admins.
cluster-admin has full access and can list all namespaces and create secrets in all namespaces.
dedicated admin needs to have read and write access to work with any namespace.
if sufficient access is missing then the secret listing or creation is blocked and the input field is disabled.